+++ CHANGELOG.md
@@ -7,9 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `operator` and `misspelled` to the `productSuggestions` query
+
 ## [1.26.0] - 2020-11-13
 
-## Added
+### Added
 
 - X-VTEX-IS-ID header to calls to search.biggylabs